+++ frontend/app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         tools:targetApi="31">
         <meta-data android:name="com.kakao.sdk.AppKey" android:value="807ed9849aecb37184efe54c84189e3b"/>
         <activity
-            android:name=".MainActivity"
+            android:name=".activity.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MJUSubwayStation_FE.NoActionBar">
