+++ frontend/app/src/main/java/com/example/mjusubwaystation_fe/activity/MainActivity.java
@@ -1,0 +1,172 @@
+package com.example.mjusubwaystation_fe.activity;
+import static android.content.ContentValues.TAG;
+
+import android.Manifest;
+
+import android.content.pm.PackageManager;
+import android.os.Build;
+import android.os.Bundle;
+
+import com.example.mjusubwaystation_fe.R;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.util.Log;
+import android.view.View;
+
+import androidx.core.content.ContextCompat;
+
+import android.widget.TextView;
+import android.widget.Button;
+import android.widget.EditText;
+
+import com.example.mjusubwaystation_fe.service.RetrofitInterface;
+import com.example.mjusubwaystation_fe.service.TestDTO;
+import com.github.chrisbanes.photoview.PhotoView;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
+public class MainActivity extends AppCompatActivity {
+    private TextView output;
+    public EditText startpoint_input, destination_input;
+    public Button find_path;
+    public static String startpoint, destination;
+    public static Call<TestDTO> call;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        PhotoView photoView = findViewById(R.id.photoView);
+        photoView.setImageResource(R.drawable.image4);
+
+        Retrofit retrofit = new Retrofit.Builder()
+                .baseUrl("https://jsonplaceholder.typicode.com/")
+                .addConverterFactory(GsonConverterFactory.create())
+                .build();
+
+        RetrofitInterface service1 = retrofit.create(RetrofitInterface.class);
+
+        // 위젯에 대한 참조.
+        output = (TextView) findViewById(R.id.tv_outPut);
+        find_path = (Button) findViewById(R.id.btn_login);
+        startpoint_input = (EditText) findViewById(R.id.edit_email);
+        destination_input = (EditText) findViewById(R.id.edit_password);
+
+        Callback fun = new Callback<TestDTO>() {
+            @Override
+            public void onResponse(Call<TestDTO> call, Response<TestDTO> response) {
+                if(response.isSuccessful()){
+                    TestDTO result = response.body();
+                    Log.d(TAG, "성공 : \n" + result.toString());
+                }
+                else{
+                    Log.d(TAG, "실패 : \n");
+                }
+            }
+
+            @Override
+            public void onFailure(Call<TestDTO> call, Throwable t) {
+                Log.d(TAG, "onFailure : " + t.getMessage());
+            }
+        };
+
+        find_path.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                startpoint = startpoint_input.getText().toString();
+                destination = destination_input.getText().toString();
+                call = service1.get_data(startpoint);
+                call.enqueue(fun);
+            }
+        });
+
+
+
+        /*photoView.setOnTouchListener(new View.OnTouchListener() {
+            public boolean onTouch(View v, MotionEvent event) {
+                switch(event.getAction()){
+                    case MotionEvent.ACTION_DOWN:
+                        break;
+                        case MotionEvent.ACTION_MOVE:
+                            v.setX(event.getX());
+                            v.setY(event.getY());
+                            break;
+                            case MotionEvent.ACTION_CANCEL:
+                                case MotionEvent.ACTION_UP:
+                                    break;
+                }
+                return true;
+            }
+
+
+        });*/
+
+        /*binding = ActivityMainBinding.inflate(getLayoutInflater());
+        setContentView(binding.getRoot());
+
+        setSupportActionBar(binding.toolbar);
+
+        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_content_main);
+        appBarConfiguration = new AppBarConfiguration.Builder(navController.getGraph()).build();
+        NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
+
+        binding.fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
+                        .setAction("Action", null).show();
+            }
+        });*/
+
+        int permission = ContextCompat.checkSelfPermission(this,
+                Manifest.permission.INTERNET);
+
+        int permission2 = ContextCompat.checkSelfPermission(this,
+                Manifest.permission.ACCESS_FINE_LOCATION);
+
+        int permission3 = ContextCompat.checkSelfPermission(this,
+                Manifest.permission.ACCESS_COARSE_LOCATION);
+
+        // 권한이 열려있는지 확인
+        if (permission == PackageManager.PERMISSION_DENIED || permission2 == PackageManager.PERMISSION_DENIED || permission3 == PackageManager.PERMISSION_DENIED) {
+            // 마쉬멜로우 이상버전부터 권한을 물어본다
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                // 권한 체크(READ_PHONE_STATE의 requestCode를 1000으로 세팅
+                requestPermissions(
+                        new String[]{Manifest.permission.INTERNET, Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION},
+                        1000);
+            }
+            return;
+        }
+    }
+
+    // 권한 체크 이후 로직
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grandResults) {
+        // READ_PHONE_STATE의 권한 체크 결과를 불러온다
+        super.onRequestPermissionsResult(requestCode, permissions, grandResults);
+        if (requestCode == 1000) {
+            boolean check_result = true;
+
+            // 모든 퍼미션을 허용했는지 체크
+            for (int result : grandResults) {
+                if (result != PackageManager.PERMISSION_GRANTED) {
+                    check_result = false;
+                    break;
+                }
+            }
+
+            // 권한 체크에 동의를 하지 않으면 안드로이드 종료
+            if (check_result == false) {
+                finish();
+            }
+        }
+
+    }
+}
