+++ frontend/app/src/main/java/com/example/mjusubwaystation_fe/fragment/SecondFragment.java
@@ -1,4 +1,4 @@
-package com.example.mjusubwaystation_fe;
+package com.example.mjusubwaystation_fe.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,11 +9,12 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.fragment.NavHostFragment;
 
-import com.example.mjusubwaystation_fe.databinding.FragmentFirstBinding;
+import com.example.mjusubwaystation_fe.R;
+import com.example.mjusubwaystation_fe.databinding.FragmentSecondBinding;
 
-public class FirstFragment extends Fragment {
+public class SecondFragment extends Fragment {
 
-    private FragmentFirstBinding binding;
+    private FragmentSecondBinding binding;
 
     @Override
     public View onCreateView(
@@ -21,22 +22,21 @@ public class FirstFragment extends Fragment {
             Bundle savedInstanceState
     ) {
 
-        binding = FragmentFirstBinding.inflate(inflater, container, false);
+        binding = FragmentSecondBinding.inflate(inflater, container, false);
         return binding.getRoot();
 
     }
 
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
-        /*
         super.onViewCreated(view, savedInstanceState);
 
-        binding.buttonFirst.setOnClickListener(new View.OnClickListener() {
+        binding.buttonSecond.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                NavHostFragment.findNavController(FirstFragment.this)
-                        .navigate(R.id.action_FirstFragment_to_SecondFragment);
+                NavHostFragment.findNavController(SecondFragment.this)
+                        .navigate(R.id.action_SecondFragment_to_FirstFragment);
             }
-        });*/
+        });
     }
 
     @Override
