+++ frontend/app/src/main/res/layout/activity_main.xml
@@ -4,18 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <!--
-    <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        android:orientation="vertical" android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        <RelativeLayout
-            android:id="@+id/map_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-        ...
-    </RelativeLayout>-->
+    tools:context=".activity.MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -28,24 +17,28 @@
             android:orientation="vertical"
             android:background="#e6e6e6"
             android:padding="20dp" >
+
             <EditText
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:id="@+id/edit_email"
                 android:hint="출발지 입력"
                 />
+
             <EditText
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:id="@+id/edit_password"
                 android:hint="도착지 입력"
                 />
+
             <Button
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:text="길 찾기"
                 android:id="@+id/btn_login"
                 />
+
             <TextView
                 android:id="@+id/tv_outPut"
                 android:layout_width="wrap_content"
