+++ frontend/app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/navigation_home"
+        android:title="@string/first_fragment_label" />
+
+    <item
+        android:id="@+id/navigation_dashboard"
+        android:title="@string/second_fragment_label" />
+
+    <item
+        android:id="@+id/navigation_notifications"
+        android:title="@string/hello_second_fragment" />
+
+</menu>
